components/backend/pyproject.toml
[tool.poetry]
name = "archie"
version = "0.1.0"
description = ""
authors = [
  "f0e <7321764+f0e@users.noreply.github.com>",
  "newgan <53710132+newgan@users.noreply.github.com>",
]
readme = "readme.md"
repository = "https://github.com/f0e/archie"
classifiers = [
  "Topic :: System :: Archiving",
  "Topic :: System :: Archiving :: Backup",
  "Topic :: Multimedia :: Video",
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
yt-dlp = "^2024.7.7"
pydantic = "^2.3.0"
pyyaml = "^6.0.1"
rich = "^13.5.2"
fastapi = "^0.103.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
soundcloud-v2 = "^1.5.3"
scdl = "^2.11.0"
pymongo = "^4.8.0"

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.11"
ruff = "^0.5.1"
pre-commit = "^3.7.1"

[tool.poetry.scripts]
archie = 'archie.cli:archie'
serve = 'archie.debug_server:start'

[tool.black]
line-length = 130

[tool.ruff]
lint.ignore = [
  "E501", # line length
]

[tool.isort]
profile = "black"

[tool.mypy]
check_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
